+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 ### Updated
 
-- Removing `(FREE)`, `(free download)`-like strings from the track names.
+- Parsing
+  - `(FREE)`, `(free download)`-like strings are now removed from the track names.
+  - `[Vinyl]` is excluded from the album names.
+  - Parse cases like `Catalogue:CAT-000` from the description correctly when the space is missing.
+
 - All testcases are now pretty JSON files - this should bring more transparency around
   the adjustments that Bandcamp make in the future (once they get updated).
+- The `comments` field now includes the media description and credits.
 
 ## [0.8.0] 2021-04-20